requirements.txt
halo==0.0.31
ipython==8.14.0
keras==2.12.0
librosa==0.10.1
matplotlib==3.7.2
noisereduce==2.0.1
numpy==1.23.5
onnxruntime==1.15.1
openpyxl==3.1.2
pandas==2.0.3
Pillow==9.3.0
Pillow==10.1.0
plotly==5.17.0
pyaudio==0.2.14
pydub==0.25.1
scikit_learn==1.3.0
scipy==1.11.3
seaborn==0.13.0
sincnet_tensorflow==0.0.2
sounddevice==0.4.6
soundfile==0.12.1
tensorflow==2.12.0
tensorflow_intel==2.12.0
tensorflow_io==0.31.0
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
tqdm==4.66.1
umap_learn==0.5.3
visualkeras==0.0.2
webrtcvad==2.0.10
